traduccion a es6 con babel, de la api sin usara mongo
