+++ lib/index.js
@@ -2,6 +2,7 @@
 import restify from 'restify';
 import config from './config/config';
 import bunyan from 'bunyan';
+import userRoutes from './routes/users';
 
 const log = new bunyan({
   name       : 'apibunyan',
@@ -43,6 +44,7 @@ server.on('uncaughtException', (req, res, route, err) => {
 });
 
 //server.on('after', restify.plugins.auditLogger({ log: log}));
+userRoutes.applyRoutes(server, '/user');
 
 server.listen(config.port, function() {
   console.log(`${server.name}, listening at ${server.url}`);
